requirements.txt
requests>=2.31.0
python-dotenv>=1.0.0
apscheduler>=3.10.4
ovh>=1.2.0
mailjet-rest>=1.3.4
playwright>=1.44.0
pytest>=8.0.0
pytest-mock>=3.12.0
responses>=0.25.0
